Updated backend structure
