--- server/src/dto.js
@@ -1,8 +1,0 @@
-module.exports.userDTO = (user) => {
-    let userCopy = JSON.parse(JSON.stringify(user));
-
-    delete userCopy._id;
-    delete userCopy.password;
-
-    return userCopy;
-}
